Fix the dataWatcher casting error (#16)

+++ src/main/java/de/katzenpapst/amunra/mob/entity/EntityAlienBug.java
@@ -27,7 +27,7 @@ public class EntityAlienBug extends EntityMob implements IEntityNonOxygenBreathe
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, 0);
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     @Override
