+++ src/main/java/de/katzenpapst/amunra/mob/entity/EntityFlyingMob.java
@@ -56,7 +56,7 @@ public abstract class EntityFlyingMob extends EntityFlying implements IMob, IAnt
     @Override
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, 0);
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     abstract protected void performAttack(Entity target, double accelX, double accelY, double accelZ);
